product crud

+++ regainmobile/src/main/java/com/isproj2/regainmobile/services/RoleService.java
@@ -6,6 +6,6 @@ import com.isproj2.regainmobile.model.Role;
 
 public interface RoleService {
 
-    public Long getRoleId(String name);
+    public Integer getRoleId(String name);
 
 }
